split base sendMessageView functionality
