+++ eahub/localgroups/views.py
@@ -7,6 +7,7 @@ from django.contrib.sites.shortcuts import get_current_site
 from django.core.mail import send_mail
 from django.shortcuts import get_object_or_404, redirect
 from django.template.loader import render_to_string
+from django.urls import reverse
 from django.views.decorators.http import require_POST
 from django.views.generic import detail as detail_views
 from django.views.generic import edit as edit_views
@@ -96,15 +97,24 @@ class ReportGroupAbuseView(ReportAbuseView):
 
 
 class SendGroupMessageView(SendMessageView):
+    def form_valid(self, form):
+        recipient = LocalGroup.objects.get(slug=self.kwargs["slug"], is_public=True)
+        sender_email: dict = form.cleaned_data["your_email_address"]
+        message: dict = form.cleaned_data["your_message"]
+        send_mail(
+            f"{sender_email} sent {recipient.name} a message through the EA hub.",
+            render_to_string(
+                "emails/message_group.txt",
+                {"message": message, "group_name": recipient.name},
+            ),
+            sender_email,
+            recipient.get_all_emails(),
+        )
 
-    receiver_type = "group"
-
-    def get_recipient(self):
-        return LocalGroup.objects.get(slug=self.kwargs["slug"], is_public=True)
-
-    def get_recipient_email(self):
-        local_group = LocalGroup.objects.get(slug=self.kwargs["slug"])
-        return local_group.get_all_emails()
+        messages.success(
+            self.request, "Your message to " + recipient.name + " has been sent"
+        )
+        return redirect(reverse("group", args=([recipient.slug])))
 
 
 @login_required
